+++ h2integrate/converters/steel/steel.py
@@ -1,10 +1,6 @@
 from attrs import field, define
 
-from h2integrate.core.utilities import (
-    BaseConfig,
-    merge_shared_cost_inputs,
-    merge_shared_performance_inputs,
-)
+from h2integrate.core.utilities import BaseConfig, merge_shared_inputs
 from h2integrate.converters.steel.steel_baseclass import (
     SteelCostBaseClass,
     SteelPerformanceBaseClass,
@@ -37,7 +33,7 @@ class SteelPerformanceModel(SteelPerformanceBaseClass):
     def setup(self):
         super().setup()
         self.config = SteelPerformanceModelConfig.from_dict(
-            merge_shared_performance_inputs(self.options["tech_config"]["model_inputs"])
+            merge_shared_inputs(self.options["tech_config"]["model_inputs"], "performance")
         )
 
     def compute(self, inputs, outputs):
@@ -78,7 +74,7 @@ class SteelCostAndFinancialModel(SteelCostBaseClass):
     def setup(self):
         super().setup()
         self.config = SteelCostAndFinancialModelConfig.from_dict(
-            merge_shared_cost_inputs(self.options["tech_config"]["model_inputs"])
+            merge_shared_inputs(self.options["tech_config"]["model_inputs"], "cost")
         )
         # TODO Bring the steel cost model config and feedstock classes into new h2integrate
         self.cost_config = SteelCostModelConfig(