+++ h2integrate/storage/hydrogen/eco_storage.py
@@ -2,7 +2,7 @@ import numpy as np
 import openmdao.api as om
 from attrs import field, define
 
-from h2integrate.core.utilities import BaseConfig, merge_shared_performance_inputs
+from h2integrate.core.utilities import BaseConfig, merge_shared_inputs
 
 
 # TODO: fix import structure in future refactor
@@ -30,7 +30,7 @@ class H2Storage(om.ExplicitComponent):
 
     def setup(self):
         self.config = H2StorageModelConfig.from_dict(
-            merge_shared_performance_inputs(self.options["tech_config"]["model_inputs"])
+            merge_shared_inputs(self.options["tech_config"]["model_inputs"], "performance")
         )
         self.add_input(
             "hydrogen_input",